overhaul the HTML and CSS and JS code.
The code written before were working but not readable. Also would be hard to change
code later. For which used the 'Web Dev Simplified' way of creating tabs. For which
less html,css and js code was written. It was easier to read. Before was using 'W3 Schools'
tab. Which was both hard and more code needed to be written.

Next need to create a modal or something similar so that task and project input can be taken.