remove projects section.
Due to a whole lot of bugs and not getting dynamically created javascript elements,
had to remove the whole projects section. Later new code needs to be written for the projects section.
for now the whole tasks section is working.

still need to correct the date packet manager so that today and this week shows the respective tasks.